rename: PlayerSetting 클래스 패키지 위치 변경
현재는 오류가 없고 패키지 이동으로 오류를 찾아 보기 위해서 변경함
