+++ src/main/java/bridge/BridgeGame.java
@@ -6,6 +6,7 @@ import bridge.dto.PlayerLocationDto;
 import bridge.dto.PlayerTryCountDto;
 import bridge.exception.InputException;
 import bridge.service.PlayerCommend;
+import bridge.service.PlayerSetting;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
