+++ src/test/java/bridge/service/PlayerSettingTest.java
@@ -1,6 +1,5 @@
 package bridge.service;
 
-import bridge.PlayerSetting;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
